+++ backend/src/trades/router.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter, Depends
-from src.currencies.models import Currency, Price
+from decimal import Decimal
 from sqlmodel.ext.asyncio.session import AsyncSession
 from src.users.models import User
 from src.users.utils import get_current_user
@@ -10,26 +10,29 @@ from src.db import pg_db
 
 trade_router = APIRouter(prefix="/trades", tags=["trades"])
 
+
 @trade_router.post("/buy", response_model=Trade)
-async def buy_currency_endpoint(currency_code: str, amount: float, session: AsyncSession = Depends(pg_db.get_session), current_user: User = Depends(get_current_user)):
-	"""
-	Endpoint do zakupu waluty. Tylko za PLN, tylko dla zalogowanych użytkowników.
-	"""
-	trade = await buy_currency(current_user.id, currency_code, amount, session)
-	return trade
+async def buy_currency_endpoint(currency_code: str, amount: Decimal, session: AsyncSession = Depends(pg_db.get_session), current_user: User = Depends(get_current_user)):
+    """
+    Endpoint do zakupu waluty. Tylko za PLN, tylko dla zalogowanych użytkowników.
+    """
+    trade = await buy_currency(current_user.id, currency_code, amount, session)
+    return trade
+
 
 @trade_router.post("/sell", response_model=Trade)
-async def sell_currency_endpoint(currency_code: str, amount: float, session: AsyncSession = Depends(pg_db.get_session), current_user: User = Depends(get_current_user)):
-	"""
-	Endpoint do sprzedaży waluty. Tylko do PLN, tylko dla zalogowanych użytkowników.
-	"""
-	trade = await sell_currency(current_user.id, currency_code, amount, session)
-	return trade
+async def sell_currency_endpoint(currency_code: str, amount: Decimal, session: AsyncSession = Depends(pg_db.get_session), current_user: User = Depends(get_current_user)):
+    """
+    Endpoint do sprzedaży waluty. Tylko do PLN, tylko dla zalogowanych użytkowników.
+    """
+    trade = await sell_currency(current_user.id, currency_code, amount, session)
+    return trade
+
 
 @trade_router.get("/trades", response_model=List[Trade])
 async def get_user_trades_endpoint(session: AsyncSession = Depends(pg_db.get_session), current_user: User = Depends(get_current_user)):
-	"""
-	Endpoint do pobierania historii transakcji użytkownika. Tylko dla zalogowanych użytkowników.
-	"""
-	trades = await get_user_trades(current_user.id, session)
-	return trades
+    """
+    Endpoint do pobierania historii transakcji użytkownika. Tylko dla zalogowanych użytkowników.
+    """
+    trades = await get_user_trades(current_user.id, session)
+    return trades
